Move editor store into editor feature and page
